Refactor Opportunity loading code
