+++ pages/opportunities/index.tsx
@@ -1,5 +1,4 @@
 import type { NextPage, GetStaticProps } from "next";
-import { PortalOpportunity } from "lib/portal-types";
 import { Layout, SectionContent, Section } from "components/layout";
 import * as Typography from "components/typography";
 import OpportunityItem from "components/sections/opportunity-overview";
@@ -14,6 +13,7 @@ import { siteData } from "lib/site-data";
 import strings from "content/strings.json";
 import Select from "components/select";
 import { PortalProject } from "lib/airtable/project";
+import { PortalOpportunity } from "lib/airtable/opportunity";
 
 type PageProps = {
   opportunities: readonly PortalOpportunity[];
